+++ CHANGELOG.md
@@ -8,6 +8,10 @@
   * Add `-k/--keep-going` option to `plm-put-playlist` command to
     continue operation despite errors and display results in the form of
     "(a/b) playlist copied" and "(c/d) media files copied"
+  * Add `-e/--error-files` option to `plm-put-playlist` command to
+    write the list of failed files to a specified file when used with
+    `-k/--keep-going` option, with "P " prefix for failed playlists and
+    "M " prefix for failed media files
 * Changed
   * Improve build process
   * Remove error message when lyrics file is not found